isoBuilder.py: Adding a feature to parse arguments from a YAML file                like: "python3 isoBuilder.py -y /path/config.yaml"
It eliminates the need to provide command-line arguments.
However, if we forget to specify a particular argument in the YAML file,
we can still pass it through the command line,

such as "python3 isoBuilder.py -v 3.0 -y /path/config.yaml".

* command-line argument takes precedence over the value specified in the config.yaml file

for ex : if the value of photon-release-version in config.yaml is set to '4.0',
and you pass the command "python3 isoBuilder.py -y /path/config.yaml -v 5.0",
then the value '5.0' specified as the command-line argument will be used.

the command-line argument -v 5.0 takes precedence over the value specified in the config.yaml file. Therefore, the value '5.0' will be used instead of '4.0' for the photon-release-version in the execution of isoBuilder.py.

- I have removed the 'required=True' for 'photon-release-version'
  at line 354 and added an if condition to check the requirement
  after loading the YAML file and raise exception if its not present

- 'custom_packages_json' variable name is replaced with 'packagelist_file' as suggested by Oliver.

- User can pass packages as a dictionary in a YAML file, it will be stored
  in the "packages_list" argument. For ex-

 packages:
    - minimal
    - linux
    - linux-esx
    - .....

- If the user has passed 'packages_list' and 'packageslist_file',
  then the code will take the union set of both package lists.

- Same for "initrd_pkgs_list_file" and "initrd_pkgs"

- Added functions in commandutils.py:

	1. `write_pkg_list_file`(file_path, file, packages_list):
		It takes the contents of packages_list & return it as a json file

	2. `merge_json_files`(path, merged_file, file1, file2):
		This function reads the contents of two JSON files, merges them based
		on specific keys, & writes the merged data to a new JSON file at the specified path.
		It returns the full path of the merged file.

- Added functions in isoBuilder.py:

        1. `create_file`(path, file_name):
                It checks if the specified path exists and creates it if it doesn't.
		Then it joins path with the file name.

	2. `merge_packages_list`(path, merged_file, file1, file2):
		If `file1` is provided, this function calls `merge_json_files` to merge the JSON files.
 		Otherwise, it returns `file2` (custom packages) as the result.

	3. `validate_options`(options, initrd_path, logger):
                If options.packages_list is a dictionary, the code performs the following actions:

		   a. It creates a file named "custom_pkg_list.json" & "merged_pkgs.json" using the `create_file()`

    		   b. `options.packages_list` is written to a json file using the `write_pkg_list_file()`

		   c. Then it merge `options.packageslist_file` with `options.packages_list` using `merge_packages_list()`. The resulting merged packages list is then assigned to options.packageslist_file.

                   d. Same for "initrd_pkgs_list_file" and "initrd_pkgs"

+++ photon_installer/generate_initrd.sh
@@ -78,8 +78,9 @@ mkdir -p $INITRD/installer
 mv ${WORKINGDIR}/sample_ui.cfg ${INITRD}/installer
 mv ${WORKINGDIR}/EULA.txt ${INITRD}/installer
 mv ${WORKINGDIR}/$PACKAGE_LIST_FILE_BASE_NAME ${INITRD}/installer
-cp ${CUSTOM_PKG_LIST_FILE} ${INITRD}/installer
-
+if [ ! -f "${INITRD}/installer/$(basename "${CUSTOM_PKG_LIST_FILE}")" ]; then
+  cp ${CUSTOM_PKG_LIST_FILE} ${INITRD}/installer
+fi
 mkfifo ${INITRD}/dev/initctl
 mknod ${INITRD}/dev/ram0 b 1 0
 mknod ${INITRD}/dev/ram1 b 1 1
